+++ GUI/Main/MainDownload.cs
@@ -129,11 +129,13 @@ namespace CKAN.GUI
         private void UpdateCachedByDownloads(CkanModule? module)
         {
             var allGuiMods = ManageMods.AllGUIMods();
+            // Find all mods that share one or more download URLs with the given module, and so on
             var affectedMods =
                 module?.GetDownloadsGroup(allGuiMods.Values
                                                     .Select(guiMod => guiMod.ToModule())
                                                     .OfType<CkanModule>())
-                       .Select(other => allGuiMods[other.identifier])
+                       .Select(other => allGuiMods.GetValueOrDefault(other.identifier))
+                       .OfType<GUIMod>()
                       ?? allGuiMods.Values;
             foreach (var otherMod in affectedMods)
             {